I optimized `seqdiff`, which is the diff backend of this library. With this updated `seqdiff`, the performance of `get_alignment` is significantly improved (~100x faster than the previous version).
The result of benchmark is below:

```
get_alignments/handmade short
                        time:   [3.2897 us 3.2912 us 3.2929 us]
                        change: [-65.480% -64.948% -64.456%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high severe
get_alignments/handmade long
                        time:   [5.0612 ms 5.0621 ms 5.0632 ms]
                        change: [-95.699% -95.599% -95.497%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
Benchmarking get_alignments/identical short: Collecting 100 samples in estimated 5.0075 s (1.9M iteration                                                                                                         get_alignments/identical short
                        time:   [2.5767 us 2.5774 us 2.5782 us]
                        change: [-28.595% -27.617% -26.740%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe
get_alignments/identical long
                        time:   [253.81 us 253.87 us 253.94 us]
                        change: [-27.157% -25.871% -24.667%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  4 (4.00%) high severe
Benchmarking get_alignments/completely different short: Collecting 100 samples in estimated 5.0039 s (444                                                                                                         get_alignments/completely different short
                        time:   [11.194 us 11.198 us 11.202 us]
                        change: [-93.177% -93.139% -93.107%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) high mild
  8 (8.00%) high severe
Benchmarking get_alignments/completely different long: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 8.1s, or reduce sample count to 60.
Benchmarking get_alignments/completely different long: Collecting 100 samples in estimated 8.0751 s (100                                                                                                          get_alignments/completely different long
                        time:   [80.517 ms 80.562 ms 80.636 ms]
                        change: [-99.012% -99.002% -98.993%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
```